refactor: extract dialog-entry actions from App.svelte (#2025)
Move handleNewRack, handleDelete, handleHelp, handleAddDevice,
handleImportFromNetBox, and handleOpenYamlEditor into a new
dialog-actions.ts module mirroring app-actions.ts. Parameterless
verbs, ready to be consumed as run: targets by the M14 command
registry (#2096). Behaviour-preserving.
